+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # 0.0.3-nullsafety
 
-- Added class `LazyFunction`, representing a single-argument memoized function.
+- Added class `MemoizedFunction`, representing a single-argument memoized function.
 
 # 0.0.2-nullsafety